--- a/src/memchr/x86/mod.rs
+++ b/src/memchr/x86/mod.rs
@@ -72,7 +72,7 @@ macro_rules! unsafe_ifunc {
 
 /// When std isn't available to provide runtime CPU feature detection, or if
 /// runtime CPU feature detection has been explicitly disabled, then just
-/// call our optimized SSE2 routine directly. SSE2 is avalbale on all x86_64
+/// call our optimized SSE2 routine directly. SSE2 is available on all x86_64
 /// targets, so no CPU feature detection is necessary.
 ///
 /// # Safety