This PR adds a NEON implementation of all `mem{r}chr` functions. This PR does not add NEON support for `memmem`, as deeper changes in the code are needed for the `Vector` trait. `memmem` support could possibly be addressed in a future PR.

Since AArch64 was introduced as part of ARMv8, and since ARMv8 requires NEON support, no runtime detection mechanism is needed for AArch64.