@redzic 2022-07-08
@BurntSushi do you know how to run the entire benchmark suite for the functions in this crate only (i.e., not libc or anything else) so that they are formatted like they are in this comment: https://github.com/BurntSushi/memchr/pull/103#issuecomment-1000980153?