@redzic 2022-07-08
Here are the initial benchmark numbers:

<img width="557" alt="Screen Shot 2022-07-08 at 6 41 22 PM" src="https://user-images.githubusercontent.com/48274562/178082485-db4a64ed-9abf-4b16-b768-2f70de2aaa26.png">