@CeleritasCelery 2023-07-06
I don't believe we need a build.rs version check. If the rustc version is too low, `target_has_feature("neon")` will return false and it won't compile the neon code. I tested this with 1.58.0 (the version before NEON intrinsics were stabilized) and it worked as expected (neon code was conditionally removed).