@BurntSushi 2023-07-07
I plan to bump the MSRV to Rust 1.60 soon, so we definitely shouldn't need to worry about version sniffing here.

Also, I have an M2 mac mini in the mail so that I can actually start playing around and testing this stuff too. (And get it working for `memmem` and also Teddy in `aho-corasick`.)