@CeleritasCelery 2023-07-10
I took some time over the weekend to look into this implementation. I thought the use of pair-wise addition as a replacement for movemask was pretty clever. Also using highly generic code resulted in saving about 200 lines overall. I do have a concern (perhaps unfounded) that it will be harder to reliable optimize code that is so generic. Particularly when comparing `[T; N]` to just having multiple variables. But I don't have a lot to base that on and would need to test it out.

I did a couple of refactoring things like replacing manual loops with map.

before:
```rust
let mut nv: [uint8x16_t; N] = [vdupq_n_u8(0); N];
for i in 0..N {
    nv[i] = vdupq_n_u8(n[i]);
}
 ```

after:
```rust
let nv = n.map(|x| vdupq_n_u8(x));
```

I found an [article](https://community.arm.com/arm-community-blogs/b/infrastructure-solutions-blog/posts/porting-x86-vector-bitmask-optimizations-to-arm-neon) from arm about how to emulate SSE movemask in NEON. It even specifically mentions memchr. Their recommendation was a clever use of shrn (shift right and narrow). I implemented that and saw a notable improvement in the mem{r}chr{2,3} benchmarks. However the mem{r}chr1 benchmarks were more mixed. I saw improvements every benchmark but the `huge/common` ones. Those saw a significant 20% slowdown. I tracked this down to moving the "movemask" out of the if-chain or not. If I [precompute the masks](https://github.com/CeleritasCelery/memchr/blob/dfd28bac9bb4fea303956937a45d29251d77a81a/src/memchr/aarch64/neon.rs#L137-L161), then `huge/common` is about as fast as the pair-wise add version (the current PR). But all the other benchmarks are 5-20% worse compare to [computing the masks in the if-chains](https://github.com/CeleritasCelery/memchr/blob/eb04700f0f687e9843ffeb05a0d3369b4deea348/src/memchr/aarch64/neon.rs#L144-L167). Not sure why that would be the case. My best guess is that `huge/common` tends to have the needle at the end of a 64 byte chunk, so computing upfront will give better ILP in that case. 

## repo
my repo is located at https://github.com/CeleritasCelery/memchr/tree/aarch64-neon

@redzic let me know if you would like me to merge any of these changes into your PR.

## benchmarks

<details>
<summary>shrn version compared to current PR</summary>

```
memchr1/krate/huge/never
                        time:   [7.1352 µs 7.1471 µs 7.1603 µs]
                        thrpt:  [77.382 GiB/s 77.524 GiB/s 77.653 GiB/s]
                 change:
                        time:   [+0.6017% +0.9897% +1.3760%] (p = 0.00 < 0.05)
                        thrpt:  [-1.3574% -0.9800% -0.5981%]
                        Change within noise threshold.

memchr1/krate/huge/rare time:   [8.9661 µs 8.9915 µs 9.0193 µs]
                        thrpt:  [61.432 GiB/s 61.622 GiB/s 61.797 GiB/s]
                 change:
                        time:   [-2.9805% -2.3185% -1.6634%] (p = 0.00 < 0.05)
                        thrpt:  [+1.6915% +2.3735% +3.0720%]
                        Performance has improved.

memchr1/krate/huge/uncommon
                        time:   [91.760 µs 93.070 µs 94.153 µs]
                        thrpt:  [5.8849 GiB/s 5.9533 GiB/s 6.0383 GiB/s]
                 change:
                        time:   [-12.080% -11.047% -10.251%] (p = 0.00 < 0.05)
                        thrpt:  [+11.422% +12.419% +13.740%]
                        Performance has improved.

memchr1/krate/huge/common
                        time:   [366.63 µs 367.24 µs 368.48 µs]
                        thrpt:  [1.5037 GiB/s 1.5087 GiB/s 1.5113 GiB/s]
                 change:
                        time:   [+19.535% +20.982% +21.975%] (p = 0.00 < 0.05)
                        thrpt:  [-18.016% -17.343% -16.342%]
                        Performance has regressed.

memchr1/krate/huge/verycommon
                        time:   [678.58 µs 681.87 µs 683.33 µs]
                        thrpt:  [830.30 MiB/s 832.08 MiB/s 836.12 MiB/s]
                 change:
                        time:   [-17.752% -17.468% -17.198%] (p = 0.00 < 0.05)
                        thrpt:  [+20.770% +21.164% +21.583%]
                        Performance has improved.

memchr1/krate/small/never
                        time:   [8.6905 ns 8.6956 ns 8.7003 ns]
                        thrpt:  [71.077 GiB/s 71.116 GiB/s 71.158 GiB/s]
                 change:
                        time:   [-0.2224% +0.0539% +0.3560%] (p = 0.75 > 0.05)
                        thrpt:  [-0.3547% -0.0538% +0.2229%]
                        No change in performance detected.

memchr1/krate/small/rare
                        time:   [12.431 ns 12.449 ns 12.490 ns]
                        thrpt:  [49.510 GiB/s 49.674 GiB/s 49.747 GiB/s]
                 change:
                        time:   [-4.9763% -4.5758% -4.1543%] (p = 0.00 < 0.05)
                        thrpt:  [+4.3344% +4.7952% +5.2369%]
                        Performance has improved.

memchr1/krate/small/uncommon
                        time:   [40.018 ns 40.176 ns 40.257 ns]
                        thrpt:  [15.361 GiB/s 15.392 GiB/s 15.453 GiB/s]
                 change:
                        time:   [-5.9687% -5.5785% -5.2338%] (p = 0.00 < 0.05)
                        thrpt:  [+5.5228% +5.9081% +6.3475%]
                        Performance has improved.

memchr1/krate/small/common
                        time:   [302.06 ns 302.36 ns 302.63 ns]
                        thrpt:  [2.0434 GiB/s 2.0452 GiB/s 2.0472 GiB/s]
                 change:
                        time:   [-17.432% -17.201% -16.987%] (p = 0.00 < 0.05)
                        thrpt:  [+20.462% +20.775% +21.113%]
                        Performance has improved.

memchr1/krate/small/verycommon
                        time:   [695.12 ns 697.14 ns 701.00 ns]
                        thrpt:  [903.34 MiB/s 908.33 MiB/s 910.98 MiB/s]
                 change:
                        time:   [-19.772% -19.396% -19.089%] (p = 0.00 < 0.05)
                        thrpt:  [+23.593% +24.064% +24.645%]
                        Performance has improved.

memchr1/krate/tiny/never
                        time:   [2.4850 ns 2.4863 ns 2.4884 ns]
                        thrpt:  [25.825 GiB/s 25.846 GiB/s 25.860 GiB/s]
                 change:
                        time:   [-0.1368% +0.1276% +0.4010%] (p = 0.43 > 0.05)
                        thrpt:  [-0.3994% -0.1275% +0.1370%]
                        No change in performance detected.

memchr1/krate/tiny/rare time:   [5.9281 ns 5.9331 ns 5.9408 ns]
                        thrpt:  [10.817 GiB/s 10.831 GiB/s 10.840 GiB/s]
                 change:
                        time:   [+0.2633% +0.5676% +0.8283%] (p = 0.00 < 0.05)
                        thrpt:  [-0.8215% -0.5644% -0.2627%]
                        Change within noise threshold.

memchr1/krate/tiny/uncommon
                        time:   [17.410 ns 17.424 ns 17.438 ns]
                        thrpt:  [3.6852 GiB/s 3.6882 GiB/s 3.6911 GiB/s]
                 change:
                        time:   [-10.024% -9.7574% -9.4662%] (p = 0.00 < 0.05)
                        thrpt:  [+10.456% +10.812% +11.141%]
                        Performance has improved.

memchr1/krate/tiny/common
                        time:   [65.980 ns 66.460 ns 66.756 ns]
                        thrpt:  [985.74 MiB/s 990.13 MiB/s 997.33 MiB/s]
                 change:
                        time:   [-15.484% -15.080% -14.699%] (p = 0.00 < 0.05)
                        thrpt:  [+17.231% +17.758% +18.321%]
                        Performance has improved.

memchr1/krate/empty/never
                        time:   [77.643 ps 77.847 ps 78.147 ps]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [+0.0557% +0.4086% +0.7819%] (p = 0.04 < 0.05)
                        thrpt:  [-0.7759% -0.4069% -0.0556%]
                        Change within noise threshold.

memchr2/krate/huge/never
                        time:   [13.146 µs 13.199 µs 13.250 µs]
                        thrpt:  [41.816 GiB/s 41.980 GiB/s 42.146 GiB/s]
                 change:
                        time:   [+0.6843% +1.2828% +1.8870%] (p = 0.00 < 0.05)
                        thrpt:  [-1.8520% -1.2666% -0.6796%]
                        Change within noise threshold.

memchr2/krate/huge/rare time:   [17.969 µs 18.005 µs 18.048 µs]
                        thrpt:  [30.700 GiB/s 30.774 GiB/s 30.834 GiB/s]
                 change:
                        time:   [-4.5215% -3.7169% -3.0746%] (p = 0.00 < 0.05)
                        thrpt:  [+3.1721% +3.8604% +4.7356%]
                        Performance has improved.

memchr2/krate/huge/uncommon
                        time:   [226.09 µs 226.38 µs 226.60 µs]
                        thrpt:  [2.4451 GiB/s 2.4475 GiB/s 2.4507 GiB/s]
                 change:
                        time:   [-8.7209% -8.4946% -8.2796%] (p = 0.00 < 0.05)
                        thrpt:  [+9.0270% +9.2832% +9.5541%]
                        Performance has improved.

memchr2/krate/huge/common
                        time:   [603.96 µs 605.20 µs 607.05 µs]
                        thrpt:  [934.63 MiB/s 937.49 MiB/s 939.41 MiB/s]
                 change:
                        time:   [-17.953% -17.344% -16.626%] (p = 0.00 < 0.05)
                        thrpt:  [+19.942% +20.984% +21.881%]
                        Performance has improved.

memchr2/krate/small/never
                        time:   [15.943 ns 15.964 ns 15.987 ns]
                        thrpt:  [38.680 GiB/s 38.737 GiB/s 38.788 GiB/s]
                 change:
                        time:   [+0.0762% +0.4733% +0.8188%] (p = 0.03 < 0.05)
                        thrpt:  [-0.8121% -0.4711% -0.0761%]
                        Change within noise threshold.

memchr2/krate/small/rare
                        time:   [21.251 ns 21.270 ns 21.282 ns]
                        thrpt:  [29.057 GiB/s 29.074 GiB/s 29.099 GiB/s]
                 change:
                        time:   [-4.5114% -4.3146% -4.1562%] (p = 0.00 < 0.05)
                        thrpt:  [+4.3364% +4.5091% +4.7246%]
                        Performance has improved.

memchr2/krate/small/uncommon
                        time:   [102.99 ns 103.24 ns 103.41 ns]
                        thrpt:  [5.9803 GiB/s 5.9901 GiB/s 6.0045 GiB/s]
                 change:
                        time:   [-13.628% -13.419% -13.236%] (p = 0.00 < 0.05)
                        thrpt:  [+15.256% +15.498% +15.778%]
                        Performance has improved.

memchr2/krate/small/common
                        time:   [640.69 ns 642.01 ns 643.89 ns]
                        thrpt:  [983.46 MiB/s 986.34 MiB/s 988.37 MiB/s]
                 change:
                        time:   [-17.109% -16.824% -16.526%] (p = 0.00 < 0.05)
                        thrpt:  [+19.798% +20.228% +20.640%]
                        Performance has improved.

memchr2/krate/tiny/never
                        time:   [2.9541 ns 2.9559 ns 2.9575 ns]
                        thrpt:  [21.728 GiB/s 21.740 GiB/s 21.753 GiB/s]
                 change:
                        time:   [-5.1212% -4.9222% -4.7583%] (p = 0.00 < 0.05)
                        thrpt:  [+4.9961% +5.1771% +5.3976%]
                        Performance has improved.

memchr2/krate/tiny/rare time:   [8.0903 ns 8.1105 ns 8.1329 ns]
                        thrpt:  [7.9014 GiB/s 7.9232 GiB/s 7.9430 GiB/s]
                 change:
                        time:   [+0.0375% +0.2586% +0.4866%] (p = 0.04 < 0.05)
                        thrpt:  [-0.4843% -0.2580% -0.0374%]
                        Change within noise threshold.

memchr2/krate/tiny/uncommon
                        time:   [55.707 ns 55.758 ns 55.859 ns]
                        thrpt:  [1.1504 GiB/s 1.1525 GiB/s 1.1535 GiB/s]
                 change:
                        time:   [-10.471% -9.8814% -9.4198%] (p = 0.00 < 0.05)
                        thrpt:  [+10.399% +10.965% +11.696%]
                        Performance has improved.

memchr2/krate/empty/never
                        time:   [77.596 ps 77.764 ps 77.967 ps]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-0.0495% +0.2039% +0.4528%] (p = 0.15 > 0.05)
                        thrpt:  [-0.4508% -0.2035% +0.0495%]
                        No change in performance detected.

memchr3/krate/huge/never
                        time:   [18.871 µs 18.969 µs 19.076 µs]
                        thrpt:  [29.046 GiB/s 29.210 GiB/s 29.361 GiB/s]
                 change:
                        time:   [+0.1764% +0.4990% +0.8824%] (p = 0.01 < 0.05)
                        thrpt:  [-0.8747% -0.4965% -0.1761%]
                        Change within noise threshold.

memchr3/krate/huge/rare time:   [25.413 µs 25.490 µs 25.635 µs]
                        thrpt:  [21.614 GiB/s 21.737 GiB/s 21.803 GiB/s]
                 change:
                        time:   [-2.6224% -2.0124% -1.3902%] (p = 0.00 < 0.05)
                        thrpt:  [+1.4098% +2.0538% +2.6930%]
                        Performance has improved.

memchr3/krate/huge/uncommon
                        time:   [298.54 µs 299.10 µs 300.01 µs]
                        thrpt:  [1.8468 GiB/s 1.8525 GiB/s 1.8560 GiB/s]
                 change:
                        time:   [-11.504% -10.884% -10.301%] (p = 0.00 < 0.05)
                        thrpt:  [+11.484% +12.213% +13.000%]
                        Performance has improved.

memchr3/krate/huge/common
                        time:   [957.55 µs 959.22 µs 961.59 µs]
                        thrpt:  [590.03 MiB/s 591.49 MiB/s 592.53 MiB/s]
                 change:
                        time:   [-16.793% -16.520% -16.211%] (p = 0.00 < 0.05)
                        thrpt:  [+19.347% +19.789% +20.182%]
                        Performance has improved.

memchr3/krate/small/never
                        time:   [22.480 ns 22.522 ns 22.608 ns]
                        thrpt:  [27.353 GiB/s 27.458 GiB/s 27.509 GiB/s]
                 change:
                        time:   [+0.0713% +0.3692% +0.6395%] (p = 0.02 < 0.05)
                        thrpt:  [-0.6354% -0.3679% -0.0713%]
                        Change within noise threshold.

memchr3/krate/small/rare
                        time:   [34.185 ns 34.262 ns 34.341 ns]
                        thrpt:  [18.007 GiB/s 18.049 GiB/s 18.090 GiB/s]
                 change:
                        time:   [-9.8074% -9.5446% -9.2730%] (p = 0.00 < 0.05)
                        thrpt:  [+10.221% +10.552% +10.874%]
                        Performance has improved.

memchr3/krate/small/uncommon
                        time:   [177.10 ns 177.34 ns 177.78 ns]
                        thrpt:  [3.4784 GiB/s 3.4870 GiB/s 3.4918 GiB/s]
                 change:
                        time:   [-18.861% -18.631% -18.408%] (p = 0.00 < 0.05)
                        thrpt:  [+22.562% +22.898% +23.246%]
                        Performance has improved.

memchr3/krate/small/common
                        time:   [992.52 ns 993.72 ns 995.70 ns]
                        thrpt:  [635.97 MiB/s 637.24 MiB/s 638.01 MiB/s]
                 change:
                        time:   [-17.135% -16.762% -16.375%] (p = 0.00 < 0.05)
                        thrpt:  [+19.582% +20.137% +20.678%]
                        Performance has improved.

memchr3/krate/tiny/never
                        time:   [3.7330 ns 3.7481 ns 3.7566 ns]
                        thrpt:  [17.106 GiB/s 17.145 GiB/s 17.214 GiB/s]
                 change:
                        time:   [+0.0642% +0.3702% +0.6164%] (p = 0.03 < 0.05)
                        thrpt:  [-0.6126% -0.3688% -0.0642%]
                        Change within noise threshold.

memchr3/krate/tiny/rare time:   [11.753 ns 11.761 ns 11.769 ns]
                        thrpt:  [5.4604 GiB/s 5.4641 GiB/s 5.4676 GiB/s]
                 change:
                        time:   [-10.857% -10.590% -10.258%] (p = 0.00 < 0.05)
                        thrpt:  [+11.430% +11.844% +12.179%]
                        Performance has improved.

memchr3/krate/tiny/uncommon
                        time:   [113.89 ns 113.98 ns 114.13 ns]
                        thrpt:  [576.59 MiB/s 577.31 MiB/s 577.78 MiB/s]
                 change:
                        time:   [-13.948% -13.782% -13.623%] (p = 0.00 < 0.05)
                        thrpt:  [+15.771% +15.986% +16.208%]
                        Performance has improved.

memchr3/krate/empty/never
                        time:   [77.939 ps 78.085 ps 78.235 ps]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [+0.3346% +0.5499% +0.7718%] (p = 0.00 < 0.05)
                        thrpt:  [-0.7658% -0.5469% -0.3335%]
                        Change within noise threshold.

memrchr1/krate/huge/never
                        time:   [7.0852 µs 7.0967 µs 7.1169 µs]
                        thrpt:  [77.853 GiB/s 78.074 GiB/s 78.201 GiB/s]
                 change:
                        time:   [+0.1400% +0.3809% +0.6461%] (p = 0.01 < 0.05)
                        thrpt:  [-0.6420% -0.3794% -0.1398%]
                        Change within noise threshold.

memrchr1/krate/huge/rare
                        time:   [8.8970 µs 8.9097 µs 8.9233 µs]
                        thrpt:  [62.093 GiB/s 62.188 GiB/s 62.277 GiB/s]
                 change:
                        time:   [-3.7058% -3.4803% -3.2463%] (p = 0.00 < 0.05)
                        thrpt:  [+3.3552% +3.6058% +3.8484%]
                        Performance has improved.

memrchr1/krate/huge/uncommon
                        time:   [111.84 µs 113.54 µs 114.79 µs]
                        thrpt:  [4.8267 GiB/s 4.8801 GiB/s 4.9542 GiB/s]
                 change:
                        time:   [-2.6909% -0.9079% +1.0157%] (p = 0.38 > 0.05)
                        thrpt:  [-1.0055% +0.9162% +2.7653%]
                        No change in performance detected.

memrchr1/krate/huge/common
                        time:   [365.45 µs 365.78 µs 366.20 µs]
                        thrpt:  [1.5130 GiB/s 1.5148 GiB/s 1.5161 GiB/s]
                 change:
                        time:   [+8.8198% +9.2131% +9.6228%] (p = 0.00 < 0.05)
                        thrpt:  [-8.7781% -8.4359% -8.1049%]
                        Performance has regressed.

memrchr1/krate/huge/verycommon
                        time:   [703.45 µs 704.21 µs 705.02 µs]
                        thrpt:  [804.76 MiB/s 805.69 MiB/s 806.56 MiB/s]
                 change:
                        time:   [-23.148% -22.818% -22.572%] (p = 0.00 < 0.05)
                        thrpt:  [+29.152% +29.564% +30.120%]
                        Performance has improved.

memrchr1/krate/small/never
                        time:   [8.6986 ns 8.7039 ns 8.7097 ns]
                        thrpt:  [71.001 GiB/s 71.048 GiB/s 71.091 GiB/s]
                 change:
                        time:   [-0.1756% +0.0233% +0.2034%] (p = 0.84 > 0.05)
                        thrpt:  [-0.2030% -0.0233% +0.1759%]
                        No change in performance detected.

memrchr1/krate/small/rare
                        time:   [10.693 ns 10.709 ns 10.730 ns]
                        thrpt:  [57.631 GiB/s 57.745 GiB/s 57.834 GiB/s]
                 change:
                        time:   [-4.3723% -4.1104% -3.8912%] (p = 0.00 < 0.05)
…